requirements.txt
playwright==1.40.0
pandas==2.1.4
openai==1.3.0
python-dotenv==1.0.0

